Moves string utils to the dedicated "strings" directory
