+++ src/utils/strings/capitalize/capitalize.js
