+++ src/utils/strings/capitalize/capitalize.spec.js
@@ -1,5 +1,5 @@
 import { expect } from 'chai'
-import capitalize from './capitalize'
+import capitalize from './'
 
 test('Capitalizes the given string', () => {
   expect(capitalize('foo')).to.equal('Foo')
