+++ src/utils/strings/capitalize/index.js
