+++ src/utils/strings/parsePropName/parsePropName.js
@@ -1,7 +1,7 @@
 // @flow
-import type { TBreakpointBehavior } from '../../const/defaultOptions'
-import Layout from '../../Layout'
-import toLowerCaseFirst from '../toLowerCaseFirst/index'
+import type { TBreakpointBehavior } from '../../../const/defaultOptions'
+import Layout from '../../../Layout'
+import toLowerCaseFirst from '../toLowerCaseFirst'
 
 export type TProps = {
   [propName: string]: mixed,
