+++ src/utils/strings/sanitizeTemplateString/index.js
