+++ src/utils/strings/toDashedString/index.js
