+++ src/utils/strings/toLowerCaseFirst/toLowerCaseFirst.spec.js
@@ -1,5 +1,5 @@
 import { expect } from 'chai'
-import toLowerCaseFirst from './index'
+import toLowerCaseFirst from './'
 
 test('Transforms first letter to lowercase', () => {
   expect(toLowerCaseFirst('Foo')).to.equal('foo')
